Fix LotAdjustment endpoints

+++ ThaiDuongWarehouse/Applications/Commands/LotAdjustments/RemoveLotAdjustmentCommandHandler.cs
@@ -11,9 +11,14 @@ public class RemoveLotAdjustmentCommandHandler : IRequestHandler<RemoveLotAdjust
     public async Task<bool> Handle(RemoveLotAdjustmentCommand request, CancellationToken cancellationToken)
     {
         var lotAdjustment = await _lotAdjustmentRepository.GetAdjustmentByLotId(request.LotId);
-#pragma warning disable CS8604 // Possible null reference argument.
+
+#pragma warning disable CS8602 // Dereference of a possibly null reference.
+        if (lotAdjustment.IsConfirmed)
+        {
+            throw new Exception($"Lot adjustment with Id {request.LotId} is not allowed to delete.");
+        }
+#pragma warning restore CS8602 // Dereference of a possibly null reference.
         _lotAdjustmentRepository.RemoveAdjustment(lotAdjustment);
-#pragma warning restore CS8604 // Possible null reference argument.
 
         return await _lotAdjustmentRepository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
     }
